+++ src/helpers/DateTimeHelper.php
@@ -12,6 +12,7 @@ use craft\i18n\Locale;
 use DateInterval;
 use DateTime;
 use DateTimeImmutable;
+use DateTimeInterface;
 use DateTimeZone;
 use Exception;
 use Throwable;
@@ -79,7 +80,7 @@ class DateTimeHelper
      *      - `timezone` – A [valid PHP timezone](https://php.net/manual/en/timezones.php). If set, this will override
      *        the assumed timezone per `$assumeSystemTimeZone`.
      *
-     * @param string|int|array|DateTime|null $value The value that should be converted to a DateTime object.
+     * @param string|int|array|DateTimeInterface|null $value The value that should be converted to a DateTime object.
      * @param bool $assumeSystemTimeZone Whether it should be assumed that the value was set in the system timezone if
      * the timezone was not specified. If this is `false`, UTC will be assumed.
      * @param bool $setToSystemTimeZone Whether to set the resulting DateTime object to the system timezone.
@@ -90,6 +91,10 @@ class DateTimeHelper
     {
         if ($value instanceof DateTime) {
             return $value;
+        }
+
+        if ($value instanceof DateTimeImmutable) {
+            return DateTime::createFromImmutable($value);
         }
 
         if (!$value) {